Update context algorithm regression test with normalizer
